requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
diff-match-patch==20230430
Django==5.0.4
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-import-export==4.0.0
djangorestframework==3.15.1
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
whitenoise==6.6.0
